Move sessions from lib to ai folder
